Review updates - Rename to useQueryResultData - Add queryResultStatus as a dependency
